[level 0] Title: 두 수의 나눗셈, Time: 0.02 ms, Memory: 77.1 MB -BaekjoonHub

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.java
@@ -1,6 +1,7 @@
 class Solution {
     public int solution(int num1, int num2) {
-        int answer = (int)(num1 / (float)num2 * 1000);
+        int answer = (int) ((double) num1 / num2 * 1000);
+        // int answer = num1 * 1000 / num2;
         return answer;
     }
 }
